requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
dnspython==2.8.0
eventlet==0.40.4
Flask==3.1.3
Flask-SocketIO==5.6.1
flatbuffers==20181003210633
gpiod==2.4.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
numpy==2.4.3
onnxruntime==1.24.4
openwakeword==0.6.0
packaging==26.0
protobuf==7.34.1
PyAudio==0.2.14
pycparser==3.0
pyrplidar==0.1.2
pyserial==3.5
python-engineio==4.13.1
python-socketio==5.16.1
requests==2.33.0
rplidar-roboticia==0.9.5
scikit-learn==1.8.0
scipy==1.17.1
simple-websocket==1.1.0
srt==3.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tqdm==4.67.3
urllib3==2.6.3
vosk==0.3.45
websockets==16.0
Werkzeug==3.1.7
wsproto==1.3.2
